fix(v3): int64 JSON 同时接受 number 与 string
新 body 的 protojson 把 int64 编成字符串；JSONInt64 解两种形态，请求仍发 number。

+++ examples/gameserver/main.go
@@ -61,7 +61,7 @@ func main() {
 		must(err)
 		rep, err := cli.Verify(ctx, os.Getenv("SHIMO_TOKEN"))
 		must(err)
-		fmt.Printf("auth user_id=%d role_id=%d\n", rep.UserID, rep.RoleID)
+		fmt.Printf("auth user_id=%d role_id=%d\n", int64(rep.UserID), int64(rep.RoleID))
 	case "iap":
 		cli, err := iap.New(c)
 		must(err)
